+++ src/components/index.js
@@ -2,9 +2,8 @@ import React from 'react';
 import { render } from 'react-dom';
 import { Link } from 'react-router-dom';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
-
 import s from '../styles/app.style';
-import {Routing} from './Routing';
+import {Routing} from './Home';
 
 const language = "ENG";
 const bkgrnd = "05";
